backend/requirements.txt
pytest==7.4.3
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest-check==2.2.2
pytest-xdist==3.3.1
requests==2.31.0
